backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
python-multipart
python-jose
cryptography
fastapi-mail
pymysql
alembic
argon2-cffi
haversine
